requirements.txt

discord.py==2.3.2
flask==3.0.0
requests==2.31.0
gunicorn==21.2.0
